+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(utils): determinePrecision function added
 
 .. 1.0.0 (2022-06-22)
 .. ------------------